NBA Teams
Altered the the columns and data inserted into the database for the NBA
Teams table.
